api/src/requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
dnspython==2.2.0
fastapi==0.74.0
fastapi-csrf-protect==0.2.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.1.0
motor==2.5.1
passlib==1.7.4
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pymongo==3.12.3
python-decouple==3.6
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.5
